+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [0.3.0](https://github.com/deslicer/ai-sidekick-for-splunk/compare/v0.2.0...v0.3.0) (2025-09-09)
+
+
+### Features
+
+* Smart Installation Scripts and Enhanced Splunk Host Handling ([#15](https://github.com/deslicer/ai-sidekick-for-splunk/issues/15)) ([a67fc3a](https://github.com/deslicer/ai-sidekick-for-splunk/commit/a67fc3a365ff3ef1ae81bbdf1db2d2ab780706a7))
+
+## [0.2.0](https://github.com/deslicer/ai-sidekick-for-splunk/compare/v0.1.0...v0.2.0) (2025-09-08)
+
+
+### Features
+
+* add interactive environment validation for workshop setup ([#12](https://github.com/deslicer/ai-sidekick-for-splunk/issues/12)) ([a20b8b1](https://github.com/deslicer/ai-sidekick-for-splunk/commit/a20b8b1761695311356dbe7ff7e1903a96b91e47))
+
 ## 0.1.0 (2025-09-02)